// src/app/features/solar-panel/models/endpoints/create-solar-panel.endpoint.ts
export interface CreateSolarPanelRequest {
  category: string;
  manufacturer: string;
  model: string;
  type: string;
  serialNumber: string;

  damage: Damage;
}

interface Damage {
  hotSpots: boolean;
  microCracks: boolean;
  snailTrails: boolean;
  brokenGlass: boolean;
  additionalDetails: string;
}